Updated auth.py and login.py
Updated auth.py and login.py to account for signup and login backend
