+++ App/views/index.py
@@ -113,15 +113,3 @@ def create_adminV():
         return jsonify({'message': f"Failed to create admin"})
 
 
-@index_views.route('/login')
-def login():
-
-    return render_template('login.html')
-
-
-@index_views.route('/signup', methods=['GET', 'POST'])
-def signup():
-    if request.method == 'POST':
-        create_student(request.form['username'], request.form['password'])
-        return render_template('index.html', users=get_all_students(),get_ranking=get_ranking,display_rankings=display_rankings,competitions=get_all_competitions())
-    return render_template('signup.html')
